ci: split build workflow into individual jobs (#25)
* chore(repo): resolve clippy lint warnings

* build(repo): explicitly set bin name and path

* ci: split build workflow into check, test and lint

* ci: use existing toolchain setup action

* ci: explicitly set toolchain in test job

+++ src/template/mod.rs
@@ -20,10 +20,7 @@ const BREAKWORDS: &[&str] = &[
 
 pub fn has_copyright_notice(b: &[u8]) -> bool {
     let n = std::cmp::min(1000, b.len());
-    let lower_b: Vec<u8> = b[..n]
-        .iter()
-        .map(|&c| c.to_ascii_lowercase() as u8)
-        .collect();
+    let lower_b: Vec<u8> = b[..n].iter().map(|&c| c.to_ascii_lowercase()).collect();
 
     let bytes = BREAKWORDS.iter().map(|w| w.as_bytes());
